move area report navigation into a hook and add report links to MapSearch
