+++ apps/platform/features/_reports/shared/area-search/AreaReportSearch.tsx
@@ -1,29 +1,17 @@
 import type { KeyboardEventHandler } from 'react'
 import { useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { useSelector } from 'react-redux'
-import { useRouter } from '@tanstack/react-router'
 import cx from 'classnames'
 import type { UseComboboxStateChange } from 'downshift'
 import { useCombobox } from 'downshift'
 
-import type { Dataview } from '@globalfishingwatch/api-types'
-import type { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
 import type { OceanArea, OceanAreaLocale } from '@globalfishingwatch/ocean-areas'
 import { InputText } from '@globalfishingwatch/ui-components'
-import { DEFAULT_WORKSPACE_CATEGORY, DEFAULT_WORKSPACE_ID } from '@platform/config/map/workspaces'
 
-import { OCEAN_AREAS_DATAVIEWS } from 'data/map/dataviews'
-import { selectAllDataviews } from 'features/_map/dataviews/dataviews.slice'
-import { getDataviewInstanceFromDataview } from 'features/_map/dataviews/dataviews.utils'
-import { selectContextAreasDataviews } from 'features/_map/dataviews/selectors/dataviews.categories.selectors'
-import { selectWorkspace } from 'features/_map/workspace/workspace.selectors'
-import { ReportCategory } from 'features/_reports/reports.types'
+import { useNavigateToAreaReport } from 'features/_reports/shared/area-search/area-report.hooks'
 import { TrackCategory, trackEvent } from 'features/app/analytics.hooks'
 import { t as trans } from 'features/i18n/i18n'
 import { useOceanAreas } from 'hooks/ocean-areas'
-import { ROUTE_PATHS } from 'router/routes.utils'
-import type { QueryParams } from 'types'
 import { getEventLabel } from 'utils/analytics'
 import { formatInfoField, upperFirst } from 'utils/info'
 
@@ -41,15 +29,12 @@ const getItemLabel = (item: OceanArea | null) => {
 
 function AreaReportSearch({ className }: { className?: string }) {
   const { t, i18n } = useTranslation()
-  const router = useRouter()
   const inputRef = useRef<HTMLInputElement | null>(null)
   const [areasMatching, setAreasMatching] = useState<OceanArea[]>([])
   const [selectedItem, setSelectedItem] = useState<OceanArea | null>(null)
   const [inputSearch, setInputSearch] = useState<string>('')
-  const workspace = useSelector(selectWorkspace)
-  const contextAreasDataviews = useSelector(selectContextAreasDataviews)
-  const allDataviews = useSelector(selectAllDataviews)
   const { searchOceanAreas } = useOceanAreas()
+  const navigateToAreaReport = useNavigateToAreaReport()
 
   const updateMatchingAreas = async (inputValue: string) => {
     try {
@@ -72,86 +57,6 @@ function AreaReportSearch({ className }: { className?: string }) {
       updateMatchingAreas(inputValue as string)
     }
     setInputSearch(inputValue as string)
-  }
-
-  const navigateToAreaReport = (area: OceanArea) => {
-    const dataview: Dataview | UrlDataviewInstance | undefined =
-      contextAreasDataviews?.find((dataview) => dataview.slug?.includes(area.properties?.type)) ||
-      allDataviews.find(
-        (dataview) =>
-          OCEAN_AREAS_DATAVIEWS.includes(dataview.slug as any) &&
-          dataview.slug?.includes(area.properties?.type)
-      )
-
-    if (dataview) {
-      const datasetId = dataview.datasetsConfig?.[0]?.datasetId
-      if (datasetId) {
-        const category = workspace?.category || DEFAULT_WORKSPACE_CATEGORY
-        const workspaceId = workspace?.id || DEFAULT_WORKSPACE_ID
-        const newDataviewInstance = getDataviewInstanceFromDataview(dataview as Dataview)
-
-        const mergeDataviewInstances = (
-          dataviewInstances: QueryParams['dataviewInstances'],
-          newDataviewInstance: UrlDataviewInstance
-        ): UrlDataviewInstance[] => {
-          const prevInstances = dataviewInstances || []
-          const dataviewInstance = prevInstances.find(
-            (d: UrlDataviewInstance) => d.id === dataview?.id
-          )
-          if (dataviewInstance) {
-            return prevInstances.map((d: UrlDataviewInstance) => {
-              if (d.id === dataviewInstance.id) {
-                return {
-                  ...d,
-                  config: {
-                    ...d.config,
-                    visible: true,
-                  },
-                }
-              }
-              return d
-            })
-          }
-          return [...prevInstances, { ...newDataviewInstance, config: { visible: true } }]
-        }
-
-        if (area.properties?.type === 'port') {
-          const portId = area.properties.area != null ? String(area.properties.area) : undefined
-          router.navigate({
-            to: ROUTE_PATHS.PORT_REPORT,
-            params: { category, workspaceId, portId: portId! },
-            search: (prev: QueryParams) => ({
-              ...prev,
-              reportCategory: ReportCategory.Events,
-              portsReportName: area.properties.name,
-              portsReportCountry: area.properties.area?.toString().split('-')[0]?.toUpperCase(),
-              portsReportDatasetId: datasetId,
-              dataviewInstances: mergeDataviewInstances(
-                prev.dataviewInstances,
-                newDataviewInstance
-              ),
-            }),
-          })
-        } else {
-          const areaId = area.properties.area != null ? String(area.properties.area) : undefined
-          router.navigate({
-            to: ROUTE_PATHS.WORKSPACE_REPORT,
-            params: { category, workspaceId, datasetId, areaId: areaId! },
-            search: (prev: QueryParams) => ({
-              ...prev,
-              dataviewInstances: mergeDataviewInstances(
-                prev.dataviewInstances,
-                newDataviewInstance
-              ),
-            }),
-          })
-        }
-      } else {
-        console.warn('No datasetId found for area', area)
-      }
-    } else {
-      console.warn('No dataview found for area', area)
-    }
   }
 
   const onSelectResult = ({ selectedItem, inputValue = '' }: UseComboboxStateChange<OceanArea>) => {
